requirements.txt
matplotlib
numpy
opencv-python
#opensimplex
perlin-noise
Pillow
scikit-learn
tensorflow
umap-learn
